Move ngrams_web_dump to `bing.py`
This is part of a major refactoring of the web scraping package to allow different implementations.
